PersistentVals

Documentation for PersistentVals.

PersistentVals.AbstractPVal

PersistentVals.TreeOrder

PersistentVals.TreePVal

PersistentVals.TreePVal

PersistentVals._TimestampInvalid

Base.get

PersistentVals.children

PersistentVals.set!

PersistentVals.AbstractPVal — Type

Abstract type for a persistent value.

A persistent value can be set at a timestamp::D and recovered later using that timestamp. Updates at different timestamps may or may not change the result at other (eg future) timestamps.

Interface

Base.get(v::AbstractVal{T,D}, d::D): value at timestanp d.

set!(v::AbstractVal{T,D}, newv::T, d::D): set value to newv at timestamp d.

PersistentVals.TreeOrder — Type

A partial order under ≤ in a tree structure.

When TreePVal needs to expand a leaf, it will call the children method (see below). This means that we can "invalidate" certain histories: by ignoring certain branches of the tree, we render them impossible.

PersistentVals.TreePVal — Type

TreePVal{T,D <: TreeOrder}

Persistent value for a tree-shaped history.

Values are only stored in leaves. Trying to access the value of a non-leaf node will throw _TimestampInvalid. Trying to access a value that is outside of the tree (as defined by the children function) will also throw _TimestampInvalid.

Type parameters

T: The stored value type.

D: The timestamp type. Must be a tree-shaped partial order under <. The root must be accessible as zero(::D). children(d::D) should give the children of node d in the partial order.

PersistentVals.TreePVal — Method

TreePVal{T,D<:TreeOrder}(val::T, date::D=zero(D), children=[])

Construct a TreePVal.

Arguments

val: The value at timestamp date.

date: The current timestamp. Defaults to zero, which should be the root of D.

PersistentVals._TimestampInvalid — Type

Error thrown when trying to access an inaccessible timestamp

Base.get — Method

get(v::AbstractPVal{T,D}, d::D)

Get value of persistent data v at time d.

PersistentVals.children — Method

children(n::TreeOrder; target=nothing)

The list of children of n in the TreeOrder. Optionally, target can be used to create a subtree of the entire TreeOrder.

PersistentVals.set! — Method

set!(v::AbstractPVal{T,D}, newv::T, d::D)

Set value v to newv at time d.